Add the module kpcb
This module is here to add a _lot_ of getters and setters for
the pcb. This is mainly to reduce the coupling between all the
kprocess file, and to add an abstraction level.
To view all the getters/setters go on src/kernel/kpcb.h
The pcb structure also change:
- add the state field, wich can have this value:  READY, RUNNING, BLOCKED, SLEEPING, WAITING_IO, DOING_IO, WAITING_CHILD, OMG_ZOMBIE
- change the field wait to sleep
- add waitfor, wich is use to keep the pid we are waiting for

With this other modification are made:
- rename NSUPERVISOR to MAXPCB, now every list had a size of MAXPCB and the system can not contain more than MAXPCB process. So no lame problem like process can not move from a list to an other freely.
- the wait field has been changed in every occurence
- same for NSUPERVISOR

At this point all the current test are succefull.
Futur change will be made with an effort to keep the backward
compatibility.