Modify kprocess to support the new kpcb and list
Again this commit make a _lot_ of change.
First of all the logic between the process change, kpcb allow to
access to the content of a process, but only get or set.
kprocess_list manage a list of pcb, but only list related
function are implement (add pcb, move pcb, remove pcb, etc..).
To be able to manage both of them in an OS, kprocess play the
role of the manager. The only file to include now is kprocess.h
The function implemented in kprocess have an higher level than
kprocess_list and kpcb. Meaning that this functions are here to
manage process in an OS. (see kprocess.h for full modification)
(I don't really know if everything is clear, but hey,
DON'T PANIC!)

This commit add test for this implementation. Only modified
functions are tested.

And finally, this commit add his pack of deprecated functions and
modified function in other modules, and bugs corrections.